+++ integration_tests/suite/test_emails.py
@@ -5,6 +5,7 @@
 from hamcrest import (
     assert_that,
     contains_inanyorder,
+    empty,
     has_entries
 )
 from xivo_test_helpers.hamcrest.uuid_ import uuid_
@@ -23,7 +24,7 @@ THREE = dict(address='three@example.com', main=False, confirmed=True)
 class TestEmails(MockBackendTestCase):
 
     @fixtures.http_user(username='foobar')
-    def test_email_updates_admin(self, foobar):
+    def test_email_updates_as_admin(self, foobar):
         assert_http_error(404, self.client.users.update_emails, UNKNOWN_UUID, [], admin=True)
         assert_http_error(
             400,
@@ -49,3 +50,26 @@ class TestEmails(MockBackendTestCase):
                 has_entries(uuid=uuid_(), **THREE),
             )
         )
+
+    @fixtures.http_user(username='foobar', email_address='one@example.com')
+    def test_email_updates_as_user(self, foobar):
+        assert_http_error(404, self.client.users.update_emails, UNKNOWN_UUID, [])
+        assert_http_error(
+            400,
+            self.client.users.update_emails,
+            foobar['uuid'], [ONE, ONE]
+        )
+
+        email_uuid = foobar['emails'][0]['uuid']
+        result = self.client.users.update_emails(foobar['uuid'], [ONE, THREE])
+        assert_that(
+            result,
+            contains_inanyorder(
+                has_entries(uuid=email_uuid, **ONE),
+                has_entries(uuid=uuid_(), address=THREE['address'], main=THREE['main'], confirmed=False),
+                # Confirmed is ignored when modifying as a user                                   ^^^^^
+            )
+        )
+
+        result = self.client.users.update_emails(foobar['uuid'], [])
+        assert_that(result, empty())