+++ wazo_auth/tests/test_schemas.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: GPL-3.0+
 
 from unittest import TestCase
+from uuid import UUID
 from hamcrest import assert_that, empty, equal_to, has_entries
 from mock import ANY
 from werkzeug.datastructures import MultiDict
@@ -105,46 +106,12 @@ class TenantSchema(TestCase):
             },
         ))
 
-    def test_that_the_uuid_is_stripped_on_load(self):
-        body = {
-            'uuid': '5ac6c192-c5c3-4448-8b51-a0d701704ce9',
-            'name': 'foobar',
-        }
-
-        result = self.schema.load(body).data
-
-        assert_that(result, equal_to({
-            'name': 'foobar',
-            'contact_uuid': None,
-            'phone': None,
-            'address': {
-                'line_1': None,
-                'line_2': None,
-                'city': None,
-                'state': None,
-                'country': None,
-                'zip_code': None,
-            },
-        }))
-
     def test_that_a_null_contact_is_accepted(self):
         body = {'contact': None}
 
         result = self.schema.load(body).data
 
-        assert_that(result, equal_to({
-            'name': None,
-            'contact_uuid': None,
-            'phone': None,
-            'address': {
-                'line_1': None,
-                'line_2': None,
-                'city': None,
-                'state': None,
-                'country': None,
-                'zip_code': None,
-            },
-        }))
+        assert_that(result, has_entries(contact_uuid=None))
 
     def test_contact_uuid_fields_when_serializing(self):
         contact = 'e04f397c-0d52-4a83-aa8e-7ee374e9eed3'
@@ -153,3 +120,28 @@ class TenantSchema(TestCase):
         result = self.schema.dump(tenant).data
 
         assert_that(result, has_entries(contact=contact))
+
+    def test_uuid(self):
+        body = {'name': 'foobar'}
+
+        result = self.schema.load(body).data
+
+        assert_that(
+            result,
+            has_entries(
+                uuid=None,
+                name='foobar',
+            )
+        )
+
+        uuid_ = body['uuid'] = 'c5b146ac-a442-4d65-a087-09a5f943ca53'
+
+        result = self.schema.load(body).data
+
+        assert_that(
+            result,
+            has_entries(
+                uuid=UUID(uuid_),
+                name='foobar',
+            )
+        )